+++ src/assets/scss/components/card.scss
@@ -6,14 +6,21 @@
   flex-direction: column;
   align-self: flex-start;
 
+  &__header {
+    overflow: hidden;
+  }
+
   &__image {
+    height: 200px;
     width: 100%;
+    object-fit: contain;
   }
 
   &__content {
     padding-top: 15px;
     display: flex;
     flex-direction: column;
+    text-align: center;
   }
 
   &__price {
@@ -32,7 +39,7 @@
 
   &:hover {
     transition: 0.5s;
-    opacity: 0.75;
+    transform: scale(1.075);
   }
 }
