+++ src/assets/scss/components/cart-modal.scss
@@ -1,18 +1,18 @@
 .cart-modal {
   position: fixed;
+  top: 50%;
+  left: 50%;
+  margin-right: -50%;
+  transform: translate(-50%, -50%);
   display: none;
   max-width: 700px;
   width: 100%;
   background-color: $white;
-  margin: auto;
   padding: 25px;
   border: 1px solid $gray;
   border-radius: 15px;
   box-shadow: 0 0 30px rgba(0, 0, 0, 0.1);
   z-index: 6;
-  top: 25%;
-  left: 0;
-  right: 0;
 
   &__btn {
     position: absolute;
