+++ src/assets/scss/components/loader.scss
@@ -1,0 +1,26 @@
+.loader {
+  display: inline-block;
+  height: 80px;
+  width: 100%;
+}
+
+.loader:after {
+  content: " ";
+  display: block;
+  width: 64px;
+  height: 64px;
+  margin: 0 auto;
+  border-radius: 50%;
+  border: 6px solid $black;
+  border-color: $black transparent $black transparent;
+  animation: loader-ring 1.2s linear infinite;
+}
+
+@keyframes loader-ring {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
