+++ src/assets/scss/index.scss
@@ -1,4 +1,5 @@
 @import 'colors';
+@import 'components/loader';
 @import 'components/header';
 @import 'components/nav';
 @import 'pages/login';
