+++ src/assets/scss/pages/login.scss
@@ -1,53 +1,53 @@
 .login {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  margin-right: -50%;
+  transform: translate(-50%, -50%);
   max-width: 360px;
   width: 100%;
-  margin: 150px auto 0 auto;
   padding: 35px 25px;
   background-color: $white;
   box-shadow: 0 0 30px rgba(0, 0, 0, 0.1);
-  border-radius: 25px;
 
   &__image {
     width: 100%;
-    max-height: 250px;
+    max-height: 225px;
     object-fit: cover;
-    border-radius: 15px 15px 0 0;
   }
 
   &__title {
     text-align: center;
-    margin: 25px 0;
+    margin: 20px 0;
   }
 
   &__input-label {
     display: block;
-    margin-bottom: 10px;
   }
 
   &__input {
     width: 100%;
     padding: 15px;
     border: 1px solid $gray;
-    border-radius: 15px;
   }
 
   &__btn {
     font-size: 1rem;
     width: 100%;
     padding: 15px;
-    margin-top: 25px;
+    margin-top: 20px;
     color: $white;
     background-color: $black;
     border: 1px solid $black;
-    border-radius: 15px;
     cursor: pointer;
 
     &:hover {
       transition: 0.5s;
-      opacity: 0.75;
+      background-color: lighten($black, 35%);
     }
 
     &:disabled {
+      background-color: $black;
       opacity: 0.5;
       cursor: auto;
     }
