+++ src/components/Card/Card.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Card from '../Card';
+import Card from './Card';
 
 describe('should render Card component', () => {
   const book = {
