+++ src/components/Loader/Loader.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Loader = () => <div className="loader" />;
+
+export default Loader;
