+++ src/components/index.js
@@ -1,4 +1,5 @@
-export { default as Header } from './Header';
-export { default as Card } from './Card';
-export { default as CartItem } from './CartItem';
-export { default as PrivateRoute } from './PrivateRouter';
+export { default as Header } from './Header/Header';
+export { default as Card } from './Card/Card';
+export { default as CartItem } from './CartItem/CartItem';
+export { default as PrivateRoute } from './PrivateRouter/PrivateRouter';
+export { default as Loader } from './Loader/Loader';
