+++ src/pages/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { HashRouter as Router, Switch, Route } from 'react-router-dom';
 
-import { Cart, Catalog, Detail, Login, NotFound } from './pages';
-import { PrivateRoute } from './components';
+import { Cart, Catalog, Detail, Login, NotFound } from './index';
+import { PrivateRoute } from '../components';
 
 const App = () => (
   <div className="wrapper">
