+++ src/pages/Login/Login.spec.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import * as redux from 'react-redux';
 
-import Login from '../Login';
+import Login from './Login';
 
 const useDispatchSpy = jest.spyOn(redux, 'useDispatch');
 
@@ -17,10 +17,5 @@ describe('should render Login component', () => {
   it('should contain .container wrapper', () => {
     const wrapper = component.find('.login');
     expect(wrapper.length).toBe(1);
-  });
-
-  it('should contain .container wrapper', () => {
-    const input = component.find('.login__input');
-    input.handleHits;
   });
 });
