+++ src/pages/index.js
@@ -1,5 +1,5 @@
-export { default as Login } from './Login';
-export { default as NotFound } from './NotFound';
-export { default as Catalog } from './Catalog';
-export { default as Cart } from './Cart';
-export { default as Detail } from './Detail';
+export { default as Login } from './Login/Login';
+export { default as NotFound } from './NotFound/NotFound';
+export { default as Catalog } from './Catalog/Catalog';
+export { default as Cart } from './Cart/Cart';
+export { default as Detail } from './Detail/Detail';
